ui: batch 6 — shared formatters, mobile-collapse sidebar
Formatters extracted into `lib/format`. Sidebar, Wall, and Servers
each carried their own fmtRate / fmtMs / fmtBytes / fmtUptime /
fmtAge / fmtClock with subtle drift in punctuation and rounding —
1.2 MiB on one page sometimes read as "1.20 MiB" on another. Now a
single module owns the canonical shape.

Sidebar collapses below md. The 256-px nav consumed most of a 360-px
mobile viewport with no escape hatch; `hidden md:flex` removes it on
phones. Operators on mobile still reach every page through the Cmd+K
palette and bookmarkable URLs. A proper slide-in hamburger is the
follow-up; this is the minimal-viable mobile fallback.

No backend changes — `Бакенд не трогаем` per operator. RESET INTERNER
and SET log_level remain documented as psql-only mutations until the
admin route grows.
